+++ api/code/firewall.go
@@ -54,6 +54,14 @@ type ForwardingBlockRule struct {
 	SrcIP    string
 }
 
+type OutputBlockRule struct {
+	BaseRule
+	Protocol string
+	DstIP    string
+	DstPort  string
+	SrcIP    string
+}
+
 type CustomInterfaceRule struct {
 	BaseRule
 	Interface string
@@ -122,6 +130,7 @@ type MulticastPort struct {
 type FirewallConfig struct {
 	ForwardingRules      []ForwardingRule
 	BlockRules           []BlockRule
+	OutputBlockRules     []OutputBlockRule
 	ForwardingBlockRules []ForwardingBlockRule
 	CustomInterfaceRules []CustomInterfaceRule
 	ServicePorts         []ServicePort
@@ -129,12 +138,13 @@ type FirewallConfig struct {
 	MulticastPorts       []MulticastPort
 	PingLan              bool
 	PingWan              bool
+	SystemDNSOverride    string
 }
 
 var FirewallConfigFile = TEST_PREFIX + "/configs/base/firewall.json"
-var gFirewallConfig = FirewallConfig{[]ForwardingRule{}, []BlockRule{},
+var gFirewallConfig = FirewallConfig{[]ForwardingRule{}, []BlockRule{}, []OutputBlockRule{},
 	[]ForwardingBlockRule{}, []CustomInterfaceRule{}, []ServicePort{},
-	[]Endpoint{}, []MulticastPort{}, false, false}
+	[]Endpoint{}, []MulticastPort{}, false, false, ""}
 
 // IP -> Iface map
 var gIfaceMap = map[string]string{}
@@ -733,6 +743,20 @@ func deleteBlock(br BlockRule) error {
 	return err
 }
 
+func deleteOutputBlock(br OutputBlockRule) error {
+	cmd := exec.Command("nft", "delete", "element", "inet", "filter", "output_block", "{",
+		br.SrcIP, ".", br.DstIP, ".", br.Protocol, ":", "drop", "}")
+
+	_, err := cmd.Output()
+
+	if err != nil {
+		log.Println("failed to delete element", err)
+		log.Println(cmd)
+	}
+
+	return err
+}
+
 func deleteForwarding(f ForwardingRule) error {
 	var cmd *exec.Cmd
 	if f.DstPort == "any" {
@@ -787,6 +811,22 @@ func applyForwarding(forwarding []ForwardingRule) error {
 func applyBlocking(blockRules []BlockRule) error {
 	for _, br := range blockRules {
 		cmd := exec.Command("nft", "add", "element", "inet", "nat", "block", "{",
+			br.SrcIP, ".", br.DstIP, ".", br.Protocol, ":", "drop", "}")
+
+		_, err := cmd.Output()
+
+		if err != nil {
+			log.Println("failed to add element", err)
+			log.Println(cmd)
+		}
+	}
+
+	return nil
+}
+
+func applyOutputBlocking(blockRules []OutputBlockRule) error {
+	for _, br := range blockRules {
+		cmd := exec.Command("nft", "add", "element", "inet", "filter", "output_block", "{",
 			br.SrcIP, ".", br.DstIP, ".", br.Protocol, ":", "drop", "}")
 
 		_, err := cmd.Output()
@@ -1128,7 +1168,7 @@ func refreshDeviceTags(dev DeviceEntry) {
 	}()
 }
 
-func refreshDeviceGroupsAndPolicy(dev DeviceEntry) {
+func refreshDeviceGroupsAndPolicy(devices map[string]DeviceEntry, groups []GroupEntry, dev DeviceEntry) {
 	if dev.WGPubKey != "" {
 		//refresh wg based on WGPubKey
 		refreshWireguardDevice(dev.MAC, dev.RecentIP, dev.WGPubKey, "wg0", "", true)
@@ -1175,7 +1215,7 @@ func refreshDeviceGroupsAndPolicy(dev DeviceEntry) {
 		addVerdictMac(ipv4, dev.MAC, ifname, "ethernet_filter", "return")
 
 		//and re-add
-		populateVmapEntries(ipv4, dev.MAC, ifname, "")
+		populateVmapEntries(devices, groups, ipv4, dev.MAC, ifname, "", dev.DNSCustom)
 	}
 
 }
@@ -1547,6 +1587,8 @@ func applyFirewallRulesLocked() {
 
 	applyBlocking(gFirewallConfig.BlockRules)
 
+	applyOutputBlocking(gFirewallConfig.OutputBlockRules)
+
 	applyForwardBlocking(gFirewallConfig.ForwardingBlockRules)
 
 	applyServicePorts(gFirewallConfig.ServicePorts)
@@ -1753,6 +1795,51 @@ func blockIP(w http.ResponseWriter, r *http.Request) {
 	}
 
 	gFirewallConfig.BlockRules = append(gFirewallConfig.BlockRules, br)
+	saveFirewallRulesLocked()
+	applyFirewallRulesLocked()
+}
+
+func blockOutputIP(w http.ResponseWriter, r *http.Request) {
+	FWmtx.Lock()
+	defer FWmtx.Unlock()
+
+	br := OutputBlockRule{}
+	err := json.NewDecoder(r.Body).Decode(&br)
+	if err != nil {
+		http.Error(w, err.Error(), 400)
+		return
+	}
+
+	if br.Protocol != "tcp" && br.Protocol != "udp" {
+		http.Error(w, "Invalid protocol", 400)
+		return
+	}
+
+	if CIDRorIP(br.SrcIP) != nil {
+		http.Error(w, "Invalid SrcIP", 400)
+		return
+	}
+
+	if CIDRorIP(br.DstIP) != nil {
+		http.Error(w, "Invalid DstIP", 400)
+		return
+	}
+
+	if r.Method == http.MethodDelete {
+		for i := range gFirewallConfig.OutputBlockRules {
+			a := gFirewallConfig.OutputBlockRules[i]
+			if br == a {
+				gFirewallConfig.OutputBlockRules = append(gFirewallConfig.OutputBlockRules[:i], gFirewallConfig.OutputBlockRules[i+1:]...)
+				saveFirewallRulesLocked()
+				deleteOutputBlock(a)
+				return
+			}
+		}
+		http.Error(w, "Not found", 404)
+		return
+	}
+
+	gFirewallConfig.OutputBlockRules = append(gFirewallConfig.OutputBlockRules, br)
 	saveFirewallRulesLocked()
 	applyFirewallRulesLocked()
 }
@@ -2059,8 +2146,30 @@ func hasVerdict(IP string, Iface string, Table string) bool {
 	return err == nil
 }
 
-func addDNSVerdict(IP string, Iface string) {
+func addCustomDNSElement(IP, DNSCustom string) {
+	err := exec.Command("nft", "add", "element", "inet", "nat", "custom_dns_devices",
+		"{", IP, ":", DNSCustom, "}").Run()
+
+	if err != nil {
+		log.Println("add custom dns server failed", IP, DNSCustom, err)
+	}
+}
+
+func delCustomDNSElement(IP, DNSCustom string) {
+	err := exec.Command("nft", "delete", "element", "inet", "nat", "custom_dns_devices",
+		"{", IP, ":", DNSCustom, "}").Run()
+
+	if err != nil {
+		log.Println("remove custom dns server failed", IP, DNSCustom, err)
+	}
+}
+
+func addDNSVerdict(IP, Iface, DNSCustom string) {
 	addVerdict(IP, Iface, "dns_access")
+
+	if DNSCustom != "" {
+		addCustomDNSElement(IP, DNSCustom)
+	}
 }
 
 func addLANVerdict(IP string, Iface string) {
@@ -2182,6 +2291,14 @@ func flushVmaps(IP string, MAC string, Ifname string, vmap_names []string, match
 	mesh_downlink := ""
 	if is_mesh {
 		mesh_downlink = meshPluginDownlink()
+	}
+
+	//check for IP in custom dns entries, and remove from there
+	entries := getCustomDNSVerdictMap()
+	for _, entry := range entries {
+		if entry.srcip == IP {
+			delCustomDNSElement(entry.srcip, entry.dstip)
+		}
 	}
 
 	for _, name := range vmap_names {
@@ -2540,9 +2657,8 @@ func getRouteGatewayForTable(Table string) string {
 	return ""
 }
 
-func populateVmapEntries(IP string, MAC string, Iface string, WGPubKey string) {
+func populateVmapEntries(devices map[string]DeviceEntry, groups []GroupEntry, IP, MAC, Iface, WGPubKey, DNSCustom string) {
 
-	groups := getGroupsJson()
 	groupsDisabled := map[string]bool{}
 	serviceGroups := map[string][]string{}
 
@@ -2553,7 +2669,6 @@ func populateVmapEntries(IP string, MAC string, Iface string, WGPubKey string) {
 		}
 	}
 
-	devices := getDevicesJson()
 	val, exists := devices[MAC]
 
 	if MAC != "" {
@@ -2600,7 +2715,7 @@ func populateVmapEntries(IP string, MAC string, Iface string, WGPubKey string) {
 	for _, policy_name := range val.Policies {
 		switch policy_name {
 		case "dns":
-			addDNSVerdict(IP, Iface)
+			addDNSVerdict(IP, Iface, DNSCustom)
 		case "lan":
 			addLANVerdict(IP, Iface)
 		case "wan":
@@ -2679,7 +2794,8 @@ func updateAddr(Router string, Ifname string) {
 	exec.Command("ip", "addr", "add", Router+"/30", "dev", Ifname).Run()
 }
 
-func establishDevice(entry DeviceEntry, new_iface string, established_route_device string, routeIP string, router string) {
+func establishDevice(devices map[string]DeviceEntry, groups []GroupEntry,
+	entry DeviceEntry, new_iface, established_route_device, routeIP, router string) {
 
 	// too noisy
 	//log.Println("flushing route and vmaps ", entry.MAC, entry.RecentIP, "`", established_route_device, "`", new_iface)
@@ -2720,10 +2836,12 @@ func establishDevice(entry DeviceEntry, new_iface string, established_route_devi
 	//add this MAC and IP to the ethernet filter. wg will be a no-op
 	addVerdictMac(entry.RecentIP, entry.MAC, new_iface, "ethernet_filter", "return")
 
+	Groupsmtx.Lock()
+	defer Groupsmtx.Unlock()
 	Devicesmtx.Lock()
 	defer Devicesmtx.Unlock()
 
-	populateVmapEntries(entry.RecentIP, entry.MAC, new_iface, entry.WGPubKey)
+	populateVmapEntries(devices, groups, entry.RecentIP, entry.MAC, new_iface, entry.WGPubKey, entry.DNSCustom)
 
 	//apply the tags
 	applyEndpointRules(entry)
@@ -2790,10 +2908,13 @@ func dynamicRouteLoop() {
 		select {
 		case <-ticker.C:
 
+			Groupsmtx.Lock()
+			groups := getGroupsJson()
 			Devicesmtx.Lock()
 			devices := getDevicesJson()
-			checkDeviceExpiries(devices)
+			checkDeviceExpiries(devices, groups)
 			Devicesmtx.Unlock()
+			Groupsmtx.Unlock()
 
 			// TBD: need to handle multiple trunk ports, lan ports
 			// that a device can arrive on.
@@ -2925,7 +3046,7 @@ func dynamicRouteLoop() {
 
 				routeIPString := routeIP + "/30"
 
-				establishDevice(entry, new_iface, established_route_device, routeIPString, routerIP)
+				establishDevice(devices, groups, entry, new_iface, established_route_device, routeIPString, routerIP)
 			}
 
 			//all devices processed, now update the iface mapping
@@ -3012,6 +3133,48 @@ func updateFirewallSubnets(DNSIP string, TinyNets []string) {
 
 }
 
+func updateSystemDNSRedirectRule(targetIP string) error {
+	err := exec.Command("nft", "flush", "chain", "inet", "filter", "DNS_OUTPUT").Run()
+	if err != nil {
+		return fmt.Errorf("failed to flush chain: %v", err)
+	}
+
+	if targetIP != "" {
+		err = exec.Command("nft", "add", "rule", "inet", "filter", "DNS_OUTPUT",
+			"inet protocol { tcp, udp }", "dport 53", "ip daddr !=", targetIP, "dnat to", targetIP).Run()
+
+		if err != nil {
+			return fmt.Errorf("failed to add rule: %v", err)
+		}
+	}
+
+	return nil
+}
+
+func systemDNSOverride(w http.ResponseWriter, r *http.Request) {
+	FWmtx.Lock()
+	defer FWmtx.Unlock()
+
+	dns := ""
+	err := json.NewDecoder(r.Body).Decode(&dns)
+	if err != nil {
+		http.Error(w, err.Error(), 400)
+		return
+	}
+
+	if dns != "" {
+		ip := net.ParseIP(dns)
+		if ip == nil {
+			http.Error(w, "Invalid IP", 400)
+			return
+		}
+	}
+
+	gFirewallConfig.SystemDNSOverride = dns
+	updateSystemDNSRedirectRule(gFirewallConfig.SystemDNSOverride)
+	saveFirewallRulesLocked()
+}
+
 func initFirewallRules() {
 	SyncBaseContainer()
 
@@ -3025,6 +3188,8 @@ func initFirewallRules() {
 	Interfacesmtx.Lock()
 	interfaces := loadInterfacesConfigLocked()
 	Interfacesmtx.Unlock()
+
+	updateSystemDNSRedirectRule(gFirewallConfig.SystemDNSOverride)
 
 	applyRadioInterfaces(interfaces)